extract path from point-LIO algorithm
use point-LIO to extract path from collection1-newercollege quad easy.
Currently still needs to be correctly aligned with starting state of the
ground-truth algorithm

+++ launch_rosScript.sh
@@ -1,13 +1,20 @@
 export WORKDIR="/Users/trading/workspace/lidar-slam"
 
+if [ $# -eq 0 ]; then
+    echo "Usage: $0 <ros script>"
+    exit 1
+fi
+
 docker build . -t rosdemo && \
-docker run --rm -it \
+docker run --rm -it\
   --network rosnet \
-  -v /Users/trading/workspace/2021-07-01-10-37-38-quad-easy.bag:/2021-07-01-10-37-38-quad-easy.bag \
+  -v /Users/trading/workspace/collection1-newercollege/2021-07-01-10-37-38-quad-easy.bag:/2021-07-01-10-37-38-quad-easy.bag \
   -v $WORKDIR/catkin_ws/:/catkin_ws/ \
   -v $WORKDIR/pointLIO_foxglove.launch:/pointLIO_foxglove.launch \
-  -v $WORKDIR/a.csv:/a.csv \
   -v $WORKDIR/bags:/bags \
+  -v $WORKDIR/$1:/$1 \
+  -p 8765:8765 \
   --env ROS_MASTER_URI=http://roscore:11311 \
+  --name pointLIO \
   rosdemo \
-  rosbag record -p /tf --output-name=/bags/recorded_data.bag
+  /bin/bash $1
